Modify type "WIP" records
We care about work in progress patches only when they're starting out
"true" Otherwise, there's no need to record for every patch that it's
not a work in progress patch.

Instead, we record it when someone sends a wip patch. Or sets a patch to
WIP or unsets a patch as WIP. Never record when wip is in the default
state.